(actualizacion)

+++ server/database/migrations/2026_01_02_012430_create_movimientos.php
@@ -15,7 +15,7 @@ return new class extends Migration
     $table->id();
     $table->unsignedBigInteger('producto_id'); 
     $table->unsignedBigInteger('venta_id')->nullable();
-    $table->integer('cantidad');
+    $table->decimal('cantidad');
     $table->string('tipo'); // entrada/salida
     $table->timestamps();
